com.hz.spi.hashslot -> com.hz.internal.util.hashslot
