+++ hazelcast/src/main/java/com/hazelcast/internal/util/hashslot/HashSlotCursor8byteKey.java
@@ -14,36 +14,30 @@
  * limitations under the License.
  */
 
-package com.hazelcast.spi.hashslot;
+package com.hazelcast.internal.util.hashslot;
 
 /**
- * Cursor over assigned slots in a {@link HashSlotArray16byteKey}. Initially the cursor's location is
- * before the first map entry and the cursor is invalid.
+ * Cursor over assigned hash slots in a {@link HashSlotArray8byteKey}.
+ * Initially the cursor's location is before the first slot and the cursor is invalid.
  */
-public interface HashSlotCursor12byteKey {
+public interface HashSlotCursor8byteKey {
     /**
      * Resets the cursor to the initial state.
      */
     void reset();
 
     /**
-     * Advances to the next assigned slot.
-     * @return true if the cursor advanced. If false is returned, the cursor is now invalid.
+     * Advance to the next assigned slot.
+     * @return {@code true} if the cursor advanced. If {@code false} is returned, the cursor is now invalid.
      * @throws IllegalStateException if a previous call to advance() already returned false.
      */
     boolean advance();
 
     /**
-     * @return key part 1 of current slot.
+     * @return the key of the current slot.
      * @throws IllegalStateException if the cursor is invalid.
      */
-    long key1();
-
-    /**
-     * @return key part 2 of current slot.
-     * @throws IllegalStateException if the cursor is invalid.
-     */
-    int key2();
+    long key();
 
     /**
      * @return Address of the current slot's value block.
