+++ hazelcast/src/main/java/com/hazelcast/internal/util/hashslot/impl/HashSlotArray12byteKeyImpl.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.hazelcast.spi.impl.hashslot;
+package com.hazelcast.internal.util.hashslot.impl;
 
-import com.hazelcast.spi.hashslot.HashSlotArray12byteKey;
-import com.hazelcast.spi.hashslot.HashSlotArray8byteKey;
-import com.hazelcast.spi.hashslot.HashSlotCursor12byteKey;
+import com.hazelcast.internal.util.hashslot.HashSlotArray12byteKey;
+import com.hazelcast.internal.util.hashslot.HashSlotArray8byteKey;
+import com.hazelcast.internal.util.hashslot.HashSlotCursor12byteKey;
 import com.hazelcast.internal.memory.MemoryManager;
 
 import static com.hazelcast.nio.Bits.INT_SIZE_IN_BYTES;
