+++ hazelcast/src/main/java/com/hazelcast/internal/util/hashslot/impl/HashSlotArrayBase.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.hazelcast.spi.impl.hashslot;
+package com.hazelcast.internal.util.hashslot.impl;
 
-import com.hazelcast.spi.hashslot.HashSlotArray;
-import com.hazelcast.spi.hashslot.HashSlotCursor12byteKey;
-import com.hazelcast.spi.hashslot.HashSlotCursor16byteKey;
-import com.hazelcast.spi.hashslot.HashSlotCursor8byteKey;
+import com.hazelcast.internal.util.hashslot.HashSlotArray;
+import com.hazelcast.internal.util.hashslot.HashSlotCursor12byteKey;
+import com.hazelcast.internal.util.hashslot.HashSlotCursor16byteKey;
+import com.hazelcast.internal.util.hashslot.HashSlotCursor8byteKey;
 import com.hazelcast.internal.memory.MemoryAccessor;
 import com.hazelcast.internal.memory.MemoryAllocator;
 import com.hazelcast.internal.memory.MemoryManager;
